requirements.txt
boto3==1.26.12
botocore==1.29.12
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
